refactor(pdf-parser): replace inline conversion logic with `DoclingConversionExecutor`
Refactor `PDFConverter` to delegate conversion operations to the new `DoclingConversionExecutor`, improving modularity and readability. Remove legacy logic including `performConversion` and associated mocks. Simplify relevant test cases to align with the new executor-based workflow, ensuring test coverage remains comprehensive.
