+++ packages/pdf-parser/src/core/pdf-converter.test.ts
@@ -1,26 +1,16 @@
 import type { LoggerMethods } from '@heripo/logger';
-import type { AsyncConversionTask, DoclingAPIClient } from 'docling-sdk';
+import type { DoclingAPIClient } from 'docling-sdk';
 
-import { existsSync, rmSync } from 'node:fs';
 import { join } from 'node:path';
 import { beforeEach, describe, expect, test, vi } from 'vitest';
 
 import { ImagePdfFallbackError } from '../errors/image-pdf-fallback-error';
-import { ImageExtractor } from '../processors/image-extractor';
 import { PdfTextExtractor } from '../processors/pdf-text-extractor';
-import { downloadTaskResult } from '../utils/docling-result-downloader';
-import { LocalFileServer } from '../utils/local-file-server';
-import { renderAndUpdatePageImages } from '../utils/page-image-updater';
-import { trackTaskProgress } from '../utils/task-progress-tracker';
 import { DocumentTypeValidator } from '../validators/document-type-validator';
 import { ChunkedPDFConverter } from './chunked-pdf-converter';
-import { buildConversionOptions } from './conversion-options-builder';
+import { DoclingConversionExecutor } from './docling-conversion-executor';
 import { ImagePdfConverter } from './image-pdf-converter';
 import { PDFConverter } from './pdf-converter';
-
-vi.mock('./conversion-options-builder', () => ({
-  buildConversionOptions: vi.fn(),
-}));
 
 vi.mock('./chunked-pdf-converter', () => ({
   ChunkedPDFConverter: vi.fn(),
@@ -30,8 +20,8 @@ vi.mock('./image-pdf-converter', () => ({
   ImagePdfConverter: vi.fn(),
 }));
 
-vi.mock('../utils/local-file-server', () => ({
-  LocalFileServer: vi.fn(),
+vi.mock('./docling-conversion-executor', () => ({
+  DoclingConversionExecutor: vi.fn(),
 }));
 
 vi.mock('node:fs', () => ({
@@ -42,16 +32,6 @@ vi.mock('node:fs', () => ({
 
 vi.mock('node:path', () => ({
   join: vi.fn(),
-}));
-
-vi.mock('../utils/docling-result-downloader', () => ({
-  downloadTaskResult: vi.fn(),
-}));
-
-vi.mock('../processors/image-extractor', () => ({
-  ImageExtractor: {
-    extractAndSaveDocumentsFromZip: vi.fn(),
-  },
 }));
 
 vi.mock('../processors/pdf-text-extractor', () => ({
@@ -66,18 +46,23 @@ vi.mock('../utils/jq', () => ({
   runJqFileJson: vi.fn(),
 }));
 
-vi.mock('../utils/page-image-updater', () => ({
-  renderAndUpdatePageImages: vi.fn(),
+vi.mock('../processors/vlm-text-corrector', () => ({
+  VlmTextCorrector: vi.fn(),
 }));
 
-vi.mock('../utils/task-progress-tracker', () => ({
-  trackTaskProgress: vi.fn(),
+vi.mock('../samplers/ocr-strategy-sampler', () => ({
+  OcrStrategySampler: vi.fn(),
+}));
+
+vi.mock('../processors/page-renderer', () => ({
+  PageRenderer: vi.fn(),
 }));
 
 describe('PDFConverter', () => {
   let logger: LoggerMethods;
   let client: DoclingAPIClient;
   let converter: PDFConverter;
+  let mockExecute: ReturnType<typeof vi.fn>;
 
   beforeEach(() => {
     logger = {
@@ -95,36 +80,12 @@ describe('PDFConverter', () => {
 
     converter = new PDFConverter(logger, client);
 
-    vi.spyOn(process, 'cwd').mockReturnValue('/test/cwd');
-    vi.spyOn(Date, 'now').mockReturnValue(1000000);
-
-    vi.mocked(buildConversionOptions).mockReturnValue({
-      to_formats: ['json', 'html'],
-      image_export_mode: 'embedded',
-      ocr_engine: 'ocrmac',
-      ocr_options: {
-        kind: 'ocrmac',
-        lang: ['ko-KR', 'en-US'],
-        recognition: 'accurate',
-        framework: 'livetext',
-      },
-      generate_picture_images: true,
-      do_picture_classification: true,
-      do_picture_description: true,
-      generate_page_images: false,
-      images_scale: 2.0,
-      force_ocr: true,
-      accelerator_options: { device: 'mps', num_threads: 4 },
-    });
-
     vi.mocked(join).mockImplementation((...args) => args.join('/'));
 
-    // Mock LocalFileServer
-    vi.mocked(LocalFileServer).mockImplementation(function () {
-      return {
-        start: vi.fn().mockResolvedValue('http://127.0.0.1:12345/test.pdf'),
-        stop: vi.fn().mockResolvedValue(undefined),
-      } as unknown as LocalFileServer;
+    // Mock DoclingConversionExecutor
+    mockExecute = vi.fn().mockResolvedValue(null);
+    vi.mocked(DoclingConversionExecutor).mockImplementation(function () {
+      return { execute: mockExecute } as any;
     });
 
     // Mock ImagePdfConverter (used by forceImagePdf and image PDF fallback)
@@ -204,9 +165,6 @@ describe('PDFConverter', () => {
       });
 
       test('does not use chunked conversion for non-file:// URLs', async () => {
-        const mockTask = createMockTask();
-        vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-
         await converter.convert(
           'http://example.com/test.pdf',
           'report-1',
@@ -234,18 +192,8 @@ describe('PDFConverter', () => {
     });
 
     test('should validate document type for file:// URL with documentValidationModel', async () => {
-      const mockTask = createMockTask();
       const onComplete = vi.fn().mockResolvedValue(undefined);
       const mockModel = {} as any;
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync)
-        .mockReturnValueOnce(true)
-        .mockReturnValueOnce(true)
-        .mockReturnValueOnce(false);
 
       await converter.convert(
         'file:///test/doc.pdf',
@@ -263,18 +211,8 @@ describe('PDFConverter', () => {
     });
 
     test('should skip validation for non-file:// URLs', async () => {
-      const mockTask = createMockTask();
       const onComplete = vi.fn().mockResolvedValue(undefined);
       const mockModel = {} as any;
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync)
-        .mockReturnValueOnce(true)
-        .mockReturnValueOnce(true)
-        .mockReturnValueOnce(false);
 
       await converter.convert(
         'http://test.com/doc.pdf',
@@ -289,15 +227,8 @@ describe('PDFConverter', () => {
     });
 
     test('should only validate once per converter instance', async () => {
-      const mockTask = createMockTask();
       const onComplete = vi.fn().mockResolvedValue(undefined);
       const mockModel = {} as any;
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(false);
 
       // First call triggers validation
       await converter.convert(
@@ -309,8 +240,6 @@ describe('PDFConverter', () => {
       );
 
       expect(mockValidate).toHaveBeenCalledTimes(1);
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(createMockTask());
 
       // Second call skips validation (already validated)
       await converter.convert(
@@ -325,17 +254,7 @@ describe('PDFConverter', () => {
     });
 
     test('should skip validation when documentValidationModel is not set', async () => {
-      const mockTask = createMockTask();
       const onComplete = vi.fn().mockResolvedValue(undefined);
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync)
-        .mockReturnValueOnce(true)
-        .mockReturnValueOnce(true)
-        .mockReturnValueOnce(false);
 
       await converter.convert(
         'file:///test/doc.pdf',
@@ -351,23 +270,59 @@ describe('PDFConverter', () => {
   });
 
   describe('convert', () => {
-    test('should successfully convert PDF with cleanupAfterCallback=false', async () => {
-      const mockTask = createMockTask();
-      const onComplete = vi.fn().mockResolvedValue(undefined);
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync)
-        .mockReturnValueOnce(true)
-        .mockReturnValueOnce(true)
-        .mockReturnValueOnce(false);
+    test('should delegate to DoclingConversionExecutor for standard conversion', async () => {
+      const onComplete = vi.fn();
 
       await converter.convert(
         'http://test.com/doc.pdf',
         'report123',
         onComplete,
+        false,
+        {},
+      );
+
+      expect(DoclingConversionExecutor).toHaveBeenCalledWith(
+        logger,
+        client,
+        expect.any(Number),
+      );
+      expect(mockExecute).toHaveBeenCalledWith(
+        'http://test.com/doc.pdf',
+        'report123',
+        onComplete,
+        false,
+        {},
+        undefined,
+      );
+    });
+
+    test('should pass abortSignal to executor', async () => {
+      const abortController = new AbortController();
+
+      await converter.convert(
+        'http://test.com/doc.pdf',
+        'report123',
+        vi.fn(),
+        false,
+        {},
+        abortController.signal,
+      );
+
+      expect(mockExecute).toHaveBeenCalledWith(
+        'http://test.com/doc.pdf',
+        'report123',
+        expect.any(Function),
+        false,
+        {},
+        abortController.signal,
+      );
+    });
+
+    test('should log converting message', async () => {
+      await converter.convert(
+        'http://test.com/doc.pdf',
+        'report123',
+        vi.fn(),
         false,
         {},
       );
@@ -376,341 +331,17 @@ describe('PDFConverter', () => {
         '[PDFConverter] Converting:',
         'http://test.com/doc.pdf',
       );
-      expect(client.convertSourceAsync).toHaveBeenCalled();
-      expect(downloadTaskResult).toHaveBeenCalled();
-      expect(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).toHaveBeenCalledWith(
-        logger,
-        '/test/cwd/result.zip',
-        '/test/cwd/result_extracted',
-        '/test/cwd/output/report123',
-      );
-      expect(onComplete).toHaveBeenCalledWith('/test/cwd/output/report123');
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/result.zip', {
-        force: true,
-      });
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/result_extracted', {
-        recursive: true,
-        force: true,
-      });
-      expect(rmSync).not.toHaveBeenCalledWith(
-        '/test/cwd/output/report123',
-        expect.any(Object),
-      );
-      expect(logger.info).toHaveBeenCalledWith(
-        '[PDFConverter] Output preserved at:',
-        '/test/cwd/output/report123',
-      );
-    });
-
-    test('should successfully convert PDF with cleanupAfterCallback=true', async () => {
-      const mockTask = createMockTask();
-      const onComplete = vi.fn().mockResolvedValue(undefined);
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(true);
-
-      await converter.convert(
-        'http://test.com/doc.pdf',
-        'report456',
-        onComplete,
-        true,
-        {},
-      );
-
-      expect(onComplete).toHaveBeenCalledWith('/test/cwd/output/report456');
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/result.zip', {
-        force: true,
-      });
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/result_extracted', {
-        recursive: true,
-        force: true,
-      });
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/output/report456', {
-        recursive: true,
-        force: true,
-      });
-      expect(logger.info).toHaveBeenCalledWith(
-        '[PDFConverter] Cleaning up output directory:',
-        '/test/cwd/output/report456',
-      );
-    });
-
-    test('should cleanup temporary files even if callback throws error', async () => {
-      const mockTask = createMockTask();
-      const onComplete = vi.fn().mockRejectedValue(new Error('Callback error'));
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(true);
-
-      await expect(
-        converter.convert(
-          'http://test.com/doc.pdf',
-          'report789',
-          onComplete,
-          false,
-          {},
-        ),
-      ).rejects.toThrow('Callback error');
-
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/result.zip', {
-        force: true,
-      });
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/result_extracted', {
-        recursive: true,
-        force: true,
-      });
-    });
-
-    test('should handle conversion task failure', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(trackTaskProgress).mockRejectedValue(
-        new Error('Task failed: Processing failed'),
-      );
-
-      await expect(
-        converter.convert(
-          'http://test.com/doc.pdf',
-          'report-fail',
-          vi.fn(),
-          false,
-          {},
-        ),
-      ).rejects.toThrow('Task failed: Processing failed');
-
-      expect(logger.error).toHaveBeenCalledWith(
-        '[PDFConverter] Conversion failed:',
-        expect.any(Error),
-      );
-    });
-
-    test('should handle download failure', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(downloadTaskResult).mockRejectedValue(
-        new Error('Download failed'),
-      );
-
-      await expect(
-        converter.convert(
-          'http://test.com/doc.pdf',
-          'report-download-fail',
-          vi.fn(),
-          false,
-          {},
-        ),
-      ).rejects.toThrow('Download failed');
-
-      expect(logger.error).toHaveBeenCalledWith(
-        '[PDFConverter] Conversion failed:',
-        expect.any(Error),
-      );
-    });
-
-    test('should handle processing failure and cleanup', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockRejectedValue(new Error('Processing failed'));
-      vi.mocked(existsSync).mockReturnValue(true);
-
-      await expect(
-        converter.convert(
-          'http://test.com/doc.pdf',
-          'report-process-fail',
-          vi.fn(),
-          false,
-          {},
-        ),
-      ).rejects.toThrow('Processing failed');
-
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/result.zip', {
-        force: true,
-      });
-      expect(rmSync).toHaveBeenCalledWith('/test/cwd/result_extracted', {
-        recursive: true,
-        force: true,
-      });
-    });
-
-    test('should handle non-existent files during cleanup', async () => {
-      const mockTask = createMockTask();
-      const onComplete = vi.fn().mockResolvedValue(undefined);
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(false);
-
-      await converter.convert(
-        'http://test.com/doc.pdf',
-        'report-no-files',
-        onComplete,
-        true,
-        {},
-      );
-
-      expect(rmSync).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('startConversionTask', () => {
-    test('should start conversion task and log task ID', async () => {
-      const mockTask = createMockTask();
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(existsSync).mockReturnValue(false);
-
-      await converter.convert(
-        'http://test.com/doc.pdf',
-        'report123',
-        vi.fn(),
-        false,
-        {},
-      );
-
-      expect(logger.info).toHaveBeenCalledWith(
-        '[PDFConverter] Task created: task-123',
-      );
-      expect(logger.info).toHaveBeenCalledWith(
-        '[PDFConverter] Polling for progress...',
-      );
-    });
-  });
-
-  describe('trackTaskProgress delegation', () => {
-    test('should call trackTaskProgress with showDetailedProgress', async () => {
-      const mockTask = createMockTask();
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(existsSync).mockReturnValue(false);
-
-      await converter.convert(
-        'http://test.com/doc.pdf',
-        'report123',
-        vi.fn(),
-        false,
-        {},
-      );
-
-      expect(trackTaskProgress).toHaveBeenCalledWith(
-        mockTask,
-        expect.any(Number),
-        logger,
-        '[PDFConverter]',
-        { showDetailedProgress: true },
-      );
-    });
-  });
-
-  describe('processConvertedFiles', () => {
-    test('should call ImageExtractor with correct paths', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(false);
-
-      await converter.convert(
-        'http://test.com/doc.pdf',
-        'report123',
-        vi.fn(),
-        false,
-        {},
-      );
-
-      expect(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).toHaveBeenCalledWith(
-        logger,
-        '/test/cwd/result.zip',
-        '/test/cwd/result_extracted',
-        '/test/cwd/output/report123',
-      );
     });
   });
 
   describe('abort signal handling', () => {
-    test('should throw AbortError when aborted after docling task completes', async () => {
-      const abortController = new AbortController();
-      const mockTask = createMockTask();
-
-      // Simulate trackTaskProgress completing but abort being triggered during it
-      vi.mocked(trackTaskProgress).mockImplementation(async () => {
-        abortController.abort();
-      });
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-
-      await expect(
-        converter.convert(
-          'http://test.com/doc.pdf',
-          'report-abort-docling',
-          vi.fn(),
-          false,
-          {},
-          abortController.signal,
-        ),
-      ).rejects.toThrow('PDF conversion was aborted');
-
-      expect(logger.info).toHaveBeenCalledWith(
-        '[PDFConverter] Conversion aborted after docling completion',
-      );
-    });
-
-    test('should throw AbortError when aborted before callback', async () => {
-      const mockTask = createMockTask();
-      const abortController = new AbortController();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-
-      // Abort after processConvertedFiles (ImageExtractor) completes
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockImplementation(async () => {
-        abortController.abort();
-      });
-      vi.mocked(existsSync).mockReturnValue(true);
-
-      await expect(
-        converter.convert(
-          'http://test.com/doc.pdf',
-          'report-abort-callback',
-          vi.fn(),
-          false,
-          {},
-          abortController.signal,
-        ),
-      ).rejects.toThrow('PDF conversion was aborted');
-
-      expect(logger.info).toHaveBeenCalledWith(
-        '[PDFConverter] Conversion aborted before callback',
-      );
-    });
-
     test('should not attempt fallback when aborted during initial conversion', async () => {
       const abortController = new AbortController();
       abortController.abort();
 
-      const mockTask = createMockTask();
-
-      // trackTaskProgress rejects (simulating task failure)
-      vi.mocked(trackTaskProgress).mockRejectedValue(
+      mockExecute.mockRejectedValue(
         new Error('Task failed: Processing failed'),
       );
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
 
       const converterWithFallback = new PDFConverter(logger, client, true);
 
@@ -732,10 +363,7 @@ describe('PDFConverter', () => {
 
   describe('image PDF fallback', () => {
     test('should not attempt fallback when enableImagePdfFallback is false', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(trackTaskProgress).mockRejectedValue(
+      mockExecute.mockRejectedValue(
         new Error('Task failed: Processing failed'),
       );
 
@@ -755,13 +383,10 @@ describe('PDFConverter', () => {
     });
 
     test('should attempt fallback when enableImagePdfFallback is true and original fails', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      // First call fails, second succeeds
-      vi.mocked(trackTaskProgress)
+      // First executor call fails, second succeeds
+      mockExecute
         .mockRejectedValueOnce(new Error('Task failed: Processing failed'))
-        .mockResolvedValueOnce(undefined);
+        .mockResolvedValueOnce(null);
 
       const mockImagePdfConverter = {
         convert: vi.fn().mockResolvedValue('/tmp/test-image.pdf'),
@@ -770,11 +395,6 @@ describe('PDFConverter', () => {
       vi.mocked(ImagePdfConverter).mockImplementation(function () {
         return mockImagePdfConverter as any;
       });
-
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(false);
 
       const converterWithFallback = new PDFConverter(logger, client, true);
 
@@ -803,10 +423,7 @@ describe('PDFConverter', () => {
     });
 
     test('should throw ImagePdfFallbackError when both original and fallback fail', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(trackTaskProgress).mockRejectedValue(
+      mockExecute.mockRejectedValue(
         new Error('Task failed: Processing failed'),
       );
 
@@ -839,10 +456,7 @@ describe('PDFConverter', () => {
     });
 
     test('should cleanup image PDF even when fallback conversion fails', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(trackTaskProgress).mockRejectedValue(
+      mockExecute.mockRejectedValue(
         new Error('Task failed: Processing failed'),
       );
 
@@ -873,13 +487,10 @@ describe('PDFConverter', () => {
     });
 
     test('should log success message when fallback succeeds', async () => {
-      const mockTask = createMockTask();
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      // First call fails, second succeeds
-      vi.mocked(trackTaskProgress)
+      // First executor call fails, second succeeds
+      mockExecute
         .mockRejectedValueOnce(new Error('Task failed: Processing failed'))
-        .mockResolvedValueOnce(undefined);
+        .mockResolvedValueOnce(null);
 
       const mockImagePdfConverter = {
         convert: vi.fn().mockResolvedValue('/tmp/test-image.pdf'),
@@ -888,11 +499,6 @@ describe('PDFConverter', () => {
       vi.mocked(ImagePdfConverter).mockImplementation(function () {
         return mockImagePdfConverter as any;
       });
-
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(false);
 
       const converterWithFallback = new PDFConverter(logger, client, true);
 
@@ -910,64 +516,8 @@ describe('PDFConverter', () => {
     });
   });
 
-  describe('renderPageImages', () => {
-    test('should call renderAndUpdatePageImages for file:// URLs', async () => {
-      const mockTask = createMockTask();
-      const onComplete = vi.fn().mockResolvedValue(undefined);
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(false);
-
-      await converter.convert(
-        'file:///test/doc.pdf',
-        'report-render',
-        onComplete,
-        false,
-        {},
-      );
-
-      expect(renderAndUpdatePageImages).toHaveBeenCalledWith(
-        '/test/doc.pdf',
-        '/test/cwd/output/report-render',
-        logger,
-        '[PDFConverter]',
-      );
-    });
-
-    test('should skip rendering and log warning for http:// URLs', async () => {
-      const mockTask = createMockTask();
-      const onComplete = vi.fn().mockResolvedValue(undefined);
-
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(
-        ImageExtractor.extractAndSaveDocumentsFromZip,
-      ).mockResolvedValue(undefined);
-      vi.mocked(existsSync).mockReturnValue(false);
-
-      await converter.convert(
-        'http://test.com/doc.pdf',
-        'report-skip-render',
-        onComplete,
-        false,
-        {},
-      );
-
-      expect(renderAndUpdatePageImages).not.toHaveBeenCalled();
-      expect(logger.warn).toHaveBeenCalledWith(
-        '[PDFConverter] Page image rendering skipped: only supported for local files (file:// URLs)',
-      );
-    });
-  });
-
   describe('forceImagePdf', () => {
     test('should convert via image PDF when forceImagePdf is true', async () => {
-      const mockTask = createMockTask();
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(existsSync).mockReturnValue(false);
-
       await converter.convert(
         'http://test.com/doc.pdf',
         'report-force-img',
@@ -983,10 +533,6 @@ describe('PDFConverter', () => {
     });
 
     test('should not convert via image PDF when forceImagePdf is false', async () => {
-      const mockTask = createMockTask();
-      vi.mocked(client.convertSourceAsync).mockResolvedValue(mockTask);
-      vi.mocked(existsSync).mockReturnValue(false);
-
       vi.mocked(ImagePdfConverter).mockClear();
 
       await converter.convert(
@@ -1018,24 +564,35 @@ describe('PDFConverter', () => {
         ),
       ).rejects.toThrow('convert failed');
     });
+
+    test('should delegate to DoclingConversionExecutor with image PDF URL', async () => {
+      const mockImagePdfConverter = {
+        convert: vi.fn().mockResolvedValue('/tmp/image.pdf'),
+        cleanup: vi.fn(),
+      };
+      vi.mocked(ImagePdfConverter).mockImplementation(function () {
+        return mockImagePdfConverter as any;
+      });
+
+      await converter.convert(
+        'http://test.com/doc.pdf',
+        'report-force-img',
+        vi.fn(),
+        false,
+        { forceImagePdf: true },
+      );
+
+      expect(mockExecute).toHaveBeenCalledWith(
+        'file:///tmp/image.pdf',
+        'report-force-img',
+        expect.any(Function),
+        false,
+        expect.objectContaining({ forceImagePdf: true }),
+        undefined,
+      );
+      expect(mockImagePdfConverter.cleanup).toHaveBeenCalledWith(
+        '/tmp/image.pdf',
+      );
+    });
   });
 });
-
-/**
- * Create a mock task that returns success on first poll.
- */
-function createMockTask(): AsyncConversionTask {
-  const task = {
-    taskId: 'task-123',
-    poll: vi.fn().mockResolvedValue({
-      task_id: 'task-123',
-      task_status: 'success',
-    }),
-    getResult: vi.fn().mockResolvedValue({
-      document: {},
-      status: 'success',
-      processing_time: 0,
-    }),
-  };
-  return task as unknown as AsyncConversionTask;
-}
