+++ packages/pdf-parser/src/core/pdf-converter.ts
@@ -1,11 +1,7 @@
 import type { LoggerMethods } from '@heripo/logger';
 import type { OcrStrategy, TokenUsageReport } from '@heripo/model';
 import type { LanguageModel } from 'ai';
-import type {
-  AsyncConversionTask,
-  ConversionOptions,
-  DoclingAPIClient,
-} from 'docling-sdk';
+import type { ConversionOptions, DoclingAPIClient } from 'docling-sdk';
 
 import { LLMTokenUsageAggregator } from '@heripo/shared';
 import { copyFileSync, existsSync, rmSync } from 'node:fs';
@@ -13,19 +9,14 @@ import { join } from 'node:path';
 
 import { CHUNKED_CONVERSION, PDF_CONVERTER } from '../config/constants';
 import { ImagePdfFallbackError } from '../errors/image-pdf-fallback-error';
-import { ImageExtractor } from '../processors/image-extractor';
 import { PageRenderer } from '../processors/page-renderer';
 import { PdfTextExtractor } from '../processors/pdf-text-extractor';
 import { VlmTextCorrector } from '../processors/vlm-text-corrector';
 import { OcrStrategySampler } from '../samplers/ocr-strategy-sampler';
-import { downloadTaskResult } from '../utils/docling-result-downloader';
 import { runJqFileJson } from '../utils/jq';
-import { LocalFileServer } from '../utils/local-file-server';
-import { renderAndUpdatePageImages } from '../utils/page-image-updater';
-import { trackTaskProgress } from '../utils/task-progress-tracker';
 import { DocumentTypeValidator } from '../validators/document-type-validator';
 import { ChunkedPDFConverter } from './chunked-pdf-converter';
-import { buildConversionOptions } from './conversion-options-builder';
+import { DoclingConversionExecutor } from './docling-conversion-executor';
 import { ImagePdfConverter } from './image-pdf-converter';
 
 /**
@@ -452,7 +443,12 @@ export class PDFConverter {
         localUrl,
       );
 
-      return await this.performConversion(
+      const executor = new DoclingConversionExecutor(
+        this.logger,
+        this.client,
+        this.timeout,
+      );
+      return await executor.execute(
         localUrl,
         reportId,
         onComplete,
@@ -482,7 +478,12 @@ export class PDFConverter {
     let originalError: Error | null = null;
 
     try {
-      return await this.performConversion(
+      const executor = new DoclingConversionExecutor(
+        this.logger,
+        this.client,
+        this.timeout,
+      );
+      return await executor.execute(
         url,
         reportId,
         onComplete,
@@ -516,7 +517,12 @@ export class PDFConverter {
       const localUrl = `file://${imagePdfPath}`;
       this.logger.info('[PDFConverter] Retrying with image PDF:', localUrl);
 
-      const report = await this.performConversion(
+      const fallbackExecutor = new DoclingConversionExecutor(
+        this.logger,
+        this.client,
+        this.timeout,
+      );
+      const report = await fallbackExecutor.execute(
         localUrl,
         reportId,
         onComplete,
@@ -539,192 +545,5 @@ export class PDFConverter {
         imagePdfConverter.cleanup(imagePdfPath);
       }
     }
-  }
-
-  private async performConversion(
-    url: string,
-    reportId: string,
-    onComplete: ConversionCompleteCallback,
-    cleanupAfterCallback: boolean,
-    options: PDFConvertOptions,
-    abortSignal?: AbortSignal,
-  ): Promise<TokenUsageReport | null> {
-    const startTime = Date.now();
-    const conversionOptions = buildConversionOptions(options);
-
-    this.logger.info(
-      `[PDFConverter] OCR languages: ${JSON.stringify(conversionOptions.ocr_options?.lang)}`,
-    );
-    this.logger.info(
-      '[PDFConverter] Converting document with Async Source API...',
-    );
-    this.logger.info('[PDFConverter] Server will download from URL directly');
-    this.logger.info(
-      '[PDFConverter] Results will be returned as ZIP to avoid memory limits',
-    );
-
-    // Resolve URL (start local server for file:// URLs)
-    const { httpUrl, server } = await this.resolveUrl(url);
-
-    try {
-      const task = await this.startConversionTask(httpUrl, conversionOptions);
-      await trackTaskProgress(
-        task,
-        this.timeout,
-        this.logger,
-        '[PDFConverter]',
-        {
-          showDetailedProgress: true,
-        },
-      );
-
-      // Check abort after docling task completes
-      if (abortSignal?.aborted) {
-        this.logger.info(
-          '[PDFConverter] Conversion aborted after docling completion',
-        );
-        const error = new Error('PDF conversion was aborted');
-        error.name = 'AbortError';
-        throw error;
-      }
-
-      const cwd = process.cwd();
-      const zipPath = join(cwd, 'result.zip');
-      await downloadTaskResult(
-        this.client,
-        task.taskId,
-        zipPath,
-        this.logger,
-        '[PDFConverter]',
-      );
-    } finally {
-      // Stop local file server if started
-      if (server) {
-        this.logger.info('[PDFConverter] Stopping local file server...');
-        await server.stop();
-      }
-    }
-
-    const cwd = process.cwd();
-    const zipPath = join(cwd, 'result.zip');
-    const extractDir = join(cwd, 'result_extracted');
-    const outputDir = join(cwd, 'output', reportId);
-
-    try {
-      await this.processConvertedFiles(zipPath, extractDir, outputDir);
-
-      // Render page images using ImageMagick (replaces Docling's page image generation)
-      if (url.startsWith('file://')) {
-        await renderAndUpdatePageImages(
-          url.slice(7),
-          outputDir,
-          this.logger,
-          '[PDFConverter]',
-        );
-      } else {
-        this.logger.warn(
-          '[PDFConverter] Page image rendering skipped: only supported for local files (file:// URLs)',
-        );
-      }
-
-      // Check abort before callback
-      if (abortSignal?.aborted) {
-        this.logger.info('[PDFConverter] Conversion aborted before callback');
-        const error = new Error('PDF conversion was aborted');
-        error.name = 'AbortError';
-        throw error;
-      }
-
-      // Execute callback with absolute output path
-      this.logger.info('[PDFConverter] Executing completion callback...');
-      await onComplete(outputDir);
-
-      const duration = Date.now() - startTime;
-      this.logger.info('[PDFConverter] Conversion completed successfully!');
-      this.logger.info('[PDFConverter] Total time:', duration, 'ms');
-    } finally {
-      // Clean up temporary files (always cleanup temp files)
-      this.logger.info('[PDFConverter] Cleaning up temporary files...');
-      if (existsSync(zipPath)) {
-        rmSync(zipPath, { force: true });
-      }
-      if (existsSync(extractDir)) {
-        rmSync(extractDir, { recursive: true, force: true });
-      }
-
-      // Cleanup output directory only if requested
-      if (cleanupAfterCallback) {
-        this.logger.info(
-          '[PDFConverter] Cleaning up output directory:',
-          outputDir,
-        );
-        if (existsSync(outputDir)) {
-          rmSync(outputDir, { recursive: true, force: true });
-        }
-      } else {
-        this.logger.info('[PDFConverter] Output preserved at:', outputDir);
-      }
-    }
-
-    return null;
-  }
-
-  private async startConversionTask(
-    url: string,
-    conversionOptions: ConversionOptions,
-  ): Promise<AsyncConversionTask> {
-    const task = await this.client.convertSourceAsync({
-      sources: [
-        {
-          kind: 'http',
-          url,
-        },
-      ],
-      options: conversionOptions,
-      target: {
-        kind: 'zip',
-      },
-    });
-
-    this.logger.info(`[PDFConverter] Task created: ${task.taskId}`);
-    this.logger.info('[PDFConverter] Polling for progress...');
-
-    return task;
-  }
-
-  /**
-   * Start a local file server for file:// URLs
-   *
-   * @param url URL to check (file:// or http://)
-   * @returns Object with httpUrl and optional server to stop later
-   */
-  private async resolveUrl(
-    url: string,
-  ): Promise<{ httpUrl: string; server?: LocalFileServer }> {
-    if (url.startsWith('file://')) {
-      const filePath = url.slice(7); // Remove 'file://' prefix
-      const server = new LocalFileServer();
-      const httpUrl = await server.start(filePath);
-
-      this.logger.info('[PDFConverter] Started local file server:', httpUrl);
-
-      return { httpUrl, server };
-    }
-
-    return { httpUrl: url };
-  }
-
-  private async processConvertedFiles(
-    zipPath: string,
-    extractDir: string,
-    outputDir: string,
-  ): Promise<void> {
-    // Extract and save documents with images
-    await ImageExtractor.extractAndSaveDocumentsFromZip(
-      this.logger,
-      zipPath,
-      extractDir,
-      outputDir,
-    );
   }
 }
